% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsMTS-package.R
\docType{package}
\name{fsMTS-package}
\alias{fsMTS-package}
\title{Feature selection for Multivariate Time Series}
\description{
Feature selection for Multivariate Time Series
}
\details{
Implementation of feature selection methods for multivariate time series
}
\author{
Dmitry Pavlyuk \email{Dmitry.V.Pavlyuk@gmail.com}
}
