% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsMTS-package.R
\docType{data}
\name{simulated}
\alias{simulated}
\title{Simulated data set}
\format{
A dataframe with 500 observations of a 10-dimensional time series
}
\usage{
data(simulated)
}
\description{
The \code{fsMTS} package includes the dataset \code{simulated},
containing information 10-dimensional time series with 500 observations
}
