% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsMTS-package.R
\docType{data}
\name{traffic.mini}
\alias{traffic.mini}
\title{Urban traffic (preprocessed and reduced)}
\format{
A dataframe with 144 observations of a 3-dimensional time series
}
\usage{
data(traffic.mini)
}
\description{
The dataset \code{traffic.mini} is a reduced data set from 3 sensors deployed on arterial roads for 12 hours
with 5-minute temporal aggregation (144 observations)
}
