% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsMTS-package.R
\docType{data}
\name{traffic}
\alias{traffic}
\title{Urban traffic (pre-processed)}
\format{
A dataframe with 288 observations of a 30-dimensional time series
}
\usage{
data(traffic)
}
\description{
The \code{fsMTS} package includes the dataset \code{traffic} that
contains information from 30 sensors deployed on arterial roads for one day
with 5-minute temporal aggregation (288 observations)
}
